requirements.txt
matplotlib>=3.7.0
networks==0.3.7
numpy==1.23.5
ogb>=1.3.2
pandas>=1.3.3
rdkit_pypi==2021.9.4              
scikit-learn>=1.1.2
scipy>=1.7.2
torch-geometric==2.5.3        
tqdm>=4.50.0
tensorboard-logger
markupsafe>=2.0.1
transformers>=4.43.4
ninja